backend/requirements.txt
flask == 2.0.0
uwsgi == 2.0.19.1
mysql-connector-python == 8.0.27